requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
bleach==2.1.1
Brotli==0.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
community==1.0.0b1
cryptography==1.9
cycler==0.10.0
d9t.json==0.4
dask==0.14.3
decorator==4.1.2
dlib==19.7.99
docopt==0.6.2
entrypoints==0.2.3
enum34==1.1.6
face-recognition==1.0.0
face-recognition-models==0.2.0
future==0.16.0
gdapi==0.2.3
h5py==2.7.1
html5lib==1.0b10
hypothesis==3.12.0
idlex==1.13
idna==2.5
imutils==0.4.3
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
Keras==1.2.1
langdetect==1.0.7
locket==0.2.0
MarkupSafe==1.0
MDP==3.5
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
notebook==5.2.1
numpy==1.13.0
nvector==0.5.2
nxviz==0.3.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
palettable==3.0.0
pandas==0.20.2
pandocfilters==1.4.2
parso==0.1.0
partd==0.3.8
patsy==0.4.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.2.1
polcart==2017.1.16.14.45
polyglot==16.7.4
prompt-toolkit==1.0.14
protobuf==3.5.2
psycopg2==2.7.3.1
ptpython==0.39
ptyprocess==0.5.2
py==1.4.34
pycosat==0.6.3
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.1.3
python-dateutil==2.6.1
python-google-places==1.4.1
python-resize-image==1.1.11
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
ruamel.yaml==0.15.34
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
Shapely==1.6.0
simplegeneric==0.8.1
singleton==0.1.0
six==1.10.0
sklearn==0.0
sqlite-bro==0.8.11
statsmodels==0.8.0
tensorflow==0.12.1
terminado==0.7
testpath==0.3.1
Theano==1.0.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.14.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.7
xlrd==1.1.0
